% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{Load_fly_app}
\alias{Load_fly_app}
\title{Load Fly App}
\usage{
Load_fly_app()
}
\value{
Launches the Fly Survival Tools Shiny application.
}
\description{
Initializes and launches the Fly data Shiny application.
}
\details{
This function first launches a small Shiny app that allows the user
to set or keep the current working directory, then launches the main
Fly Survival Tools application using that directory.
}
\examples{
\dontrun{
Load_fly_app()
}
}
